Add function getMACAddress
Construct allows passing of $type to function but $type is hard-coded into getAddress so there is no way to access other types of addresses (IPV6, mac, etc.) using object.    Information is correctly parsed from XML already but there is no other way to access it.   A more elegant solution alllowing the object to specify which type of address could be implemented which would undoubtedly be more useful to a larger population if anyone is intrested.

  public function getMACAddress($type = 'mac', $index = 0)
    {
        return $this->_addresses[$type][$index];
    }